fix: detect an empty timeline record body as KafkaNull
Message.getPayload() is never null — Spring forbids a null payload — so the
guard was dead code and a record published with a null value, which converts to
KafkaNull, would have reached deserialization instead of being discarded.

- share the empty-body check as MessagePayloads.isEmpty across all four
  timeline.events consumers
- drop a redundant cast and rename the test locals shadowing the restricted
  identifier 'record'

+++ src/main/java/io/github/temporalrift/game/shared/TimelineEventEnvelope.java
@@ -109,7 +109,7 @@ public record TimelineEventEnvelope(
 
     private static boolean isExactInt(Number number) {
         var asLong = number.longValue();
-        return asLong == (int) asLong && number.doubleValue() == (double) asLong;
+        return asLong == (int) asLong && number.doubleValue() == asLong;
     }
 
     private static Integer parseInt(String text) {
